create message search and message simulation sub module for state manager
